Added a configuration option for a user response file to the config file.
Added code to the test bot to load a response file (right now, a JSON document)
when the core class is instantiated.  Also added code to the message response
method to match the keywords in the response file against the message body if
nothing else matches and send a randomly chosen response.

Signed-off-by: The Doctor <drwho@virtadpt.net>