Got "post to Twitter" method working.  I had to regenerate my API keys after setting this app to read/write.
Added "list trends" method to search for trending topics in geographic
locations because that's what Twitter lets you do.

Added a stub method that tells the user how to download an archive of the
user's tweets because Twitter doesn't let you do that through the API.

Added a method to query a user's Twitter profile.

Added a method to query arbitrary volumes of a Twitter user's timeline.

Added more help text.

Signed-off-by: The Doctor <drwho@virtadpt.net>