Updated testbot.responses because it keeps getting updated.
Initial commit of the TwitterStreamListener() class, which is a component of
the "monitor Twitter for stuff going on" functionality.

Added a few more imports to TwitterBot() to support new functionality.  Added
some data structures to the class to implement that functionality.  Updated
help text.  Began sketching out the code which will monitor Twitter's stream
API for the search terms.

Signed-off-by: The Doctor <drwho@virtadpt.net>